Added new acquisition strategy for s2 processing
