+++ configs/logging.cfg
@@ -2,7 +2,7 @@
 keys=root, luigiStatus, gaipError, gaipStatus
 
 [handlers]
-keys=consoleHandler, fileHandler, errfileHandler
+keys=consoleHandler, fileHandler, errfileHandler, gaipfileHandler
 
 [formatters]
 keys=simpleFormatter
