+++ configs/luigi.cfg
@@ -30,7 +30,7 @@ base_dir = _interpolation
 
 [InterpolateCoefficients]
 
-[DEMExctraction]
+[DEMExtraction]
 
 [SlopeAndAspect]
