+++ gaip/acquisition/sensors.json
