+++ gaip/setup.py
@@ -15,7 +15,7 @@ def configuration(parent_package='', top_path=None):
 
     config = Configuration('gaip', parent_package, top_path)
     config.add_subpackage('scripts')
-    config.add_data_files('sensors.json')
+    config.add_data_files('acquisition/sensors.json')
     config.add_data_dir('spectral_response')
     config.add_extension(
         '__cast_shadow_mask',
