+++ gaip/singlefile_workflow.py
@@ -81,6 +81,7 @@ class DataStandardisation(luigi.Task):
     tle_path = luigi.Parameter(significant=False)
     rori = luigi.FloatParameter(default=0.52, significant=False)
     compression = luigi.Parameter(default='lzf', significant=False)
+    acq_parser_hint = luigi.Parameter(default=None)
 
     def output(self):
         fmt = '{scene}.gaip.h5'
@@ -96,7 +97,7 @@ class DataStandardisation(luigi.Task):
                    self.ozone_path, self.water_vapour_path, self.dem_path,
                    self.dsm_fname, self.invariant_height_fname,
                    self.modtran_exe, out_fname, self.ecmwf_path, self.rori,
-                   self.compression)
+                   self.compression, self.acq_parser_hint)
 
 
 class ARD(luigi.WrapperTask):
