+++ setup.py
@@ -63,6 +63,7 @@ setup(
     license='CC0 1.0 Universal',
     author='The gaip authors',
     maintainer='gaip developers',
+    packages=setuptools.find_packages(),
     scripts=['utils/test_satellite_solar_angles',
              'utils/test_dsm',
              'utils/test_exiting_angles',
